fix: remove need of buffer polyfill
Bundler such as webpack error out when encountering `Buffer` class.
While `Buffer` is not actually used in the browser, the bundler cannot
detect this from the `if` statement.

By using the same trick on teh `allocUnsafe` call, it fools the bundler
to not detect the usage of `Buffer`.

Without this, one needs to explicitly ask the bundler to not polyfill
`Buffer`.